refactored stream stats into separate class; expanded class to collect things like mean(interarrival)

+++ code/SummaryStore/src/main/java/com/samsung/sra/DataStore/Aggregates/SimpleCountOperator.java
@@ -26,7 +26,7 @@ public class SimpleCountOperator implements WindowOperator<Long, Long, Long, Lon
                 // the intersection of [a, b] and [p, q] is [max(a, p), min(b, q)]
                 long l = Math.max(qt0, bt0), r = Math.min(qt1, bt1);
                 assert r >= l && r - l <= bt1 - bt0;
-                return bCount * (r - l + 1) / (bt1 - bt0 + 1); // TODO: check for int overflow issues
+                return (long)((double)bCount * (r - l + 1d) / (bt1 - bt0 + 1d));
             }
         },
         HALF_BOUND {
